persona_system/requirements.txt
# AI/ML 라이브러리
openai>=1.0.0
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=2.0.0

# 한국어 처리
konlpy>=0.6.0
spacy>=3.6.0

# 데이터 처리 - 버전 호환성 고려
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0

# 벡터 데이터베이스
faiss-cpu>=1.7.0

# 백엔드 프레임워크
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6

# 유틸리티
python-dotenv>=0.19.0
pydantic>=2.0.0
loguru>=0.7.0

# 개발 도구
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0 